python/scientific_discovery/graph_reasoning/requirements.txt
# Core Python utilities
typing>=3.7.4
pathlib>=1.0.1
requests>=2.28.0
tqdm>=4.65.0
python-dateutil>=2.8.2

# Data processing and analysis
pandas>=1.5.0
numpy>=1.21.0
scipy>=1.7.0
scikit-learn>=1.0.0

# Deep Learning and NLP
torch>=2.0.0
transformers>=4.30.0
langchain>=0.0.200
openai>=1.0.0

# Graph Processing
networkx>=2.8.0
pyvis>=0.3.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.12.0
Pillow>=9.0.0

# Jupyter and Display
ipython>=8.0.0
jupyter>=1.0.0
ipywidgets>=8.0.0

# Utility packages
pyyaml>=6.0
python-dotenv>=0.20.0

# Development tools
pytest>=7.0.0
pytest-cov>=4.0.0
black>=22.0.0
flake8>=5.0.0
mypy>=1.0.0
isort>=5.0.0

# Documentation
sphinx>=5.0.0
sphinx-rtd-theme>=1.0.0
myst-parser>=1.0.0